Actualización

+++ Entrega 1_Léxico y Sintaxis/semantic_cube.py
@@ -37,7 +37,7 @@ class SemanticCube:
             },
             "/": {
                 INT: {
-                    INT: FLOAT,       # int / int = int
+                    INT: INT,         # int / int = int
                     FLOAT: FLOAT},    # int / float = float
                 
                 FLOAT: {
